Testing and command parsing

+++ color_correct.py
@@ -15,11 +15,11 @@ def stretch_pre(nimg):
     return nimg.transpose(1, 2, 0)
 
 def grey_world(nimg):
-    nimg = nimg.transpose(2, 0, 1).astype(np.uint32)
+    nimg = nimg.transpose(2, 0, 1)
     mu_g = np.average(nimg[1])
     nimg[0] = np.minimum(nimg[0]*(mu_g/np.average(nimg[0])),255)
     nimg[2] = np.minimum(nimg[2]*(mu_g/np.average(nimg[2])),255)
-    return  nimg.transpose(1, 2, 0).astype(np.uint8)
+    return  nimg.transpose(1, 2, 0)
 
 def max_white(nimg):
     if nimg.dtype==np.uint8:
